libtsm: build the maintained fork instead of the abandoned upstream tree
projects/libtsm has cloned git://people.freedesktop.org/~dvdhrm/libtsm since
2016. That tree's HEAD is b73acb4c, committed 2014-04-24, and it has not moved
since. Development continued in a fork, which is now at
https://github.com/kmscon/libtsm (v4.7.1) and is what Arch, Debian and Alpine
package as libtsm today.

The practical consequence is a global-buffer-overflow that OSS-Fuzz has
re-reported ten times over eight years without it ever being fixed:
OSV-2016-4, OSV-2016-5, OSV-2017-121, OSV-2017-157, OSV-2017-161,
OSV-2017-167, OSV-2020-1448, OSV-2021-289, OSV-2021-1373 and OSV-2024-777
are all "Global-buffer-overflow in vte_write_debug", separate records only
because inlining shuffles the second frame between do_execute, do_action,
csi_dev_attr, do_trans and send_primary_da. The bug is one line:

    static void send_primary_da(struct tsm_vte *vte)
    {
            vte_write(vte, "\e[?60;1;6;9;15c", 17);
    }

The literal is ESC + 14 characters + NUL = 16 bytes, so the hardcoded 17 reads
one past it on any input containing a Device Attributes query. It was reported
downstream in 2015 (tmux/tmux#49, freedesktop bug 91335) and fixed in the fork
in January 2018 with sizeof(str) - 1.

Changes:

* Dockerfile clones https://github.com/kmscon/libtsm, and installs meson and
  ninja instead of autotools -- the fork moved to meson.
* build.sh drives meson. LDFLAGS carries the sanitizer flags and b_lundef is
  disabled because the project declares a shared_library, which otherwise
  links with -Wl,--no-undefined against an unlinked sanitizer runtime. The
  objects meson emits are archived rather than the source list being
  duplicated here.
* libtsm_fuzzer.c: tsm_screen_draw_cb's first argument widened from uint32_t
  to uint64_t between the two trees; the callback signature follows.
* project.yaml: homepage and main_repo point at the new location.

Verified locally, x86_64/libfuzzer/ASan:

* Before, on a 3-byte input "\e[c":
    ERROR: AddressSanitizer: global-buffer-overflow, READ of size 1
    #0 vte_write_debug src/tsm/tsm-vte.c:514
    #1 do_trans        src/tsm/tsm-vte.c:1858
    #2 tsm_vte_input
    DEDUP_TOKEN: vte_write_debug--do_trans--tsm_vte_input
  which is OSV-2024-777's crash state exactly.
* After: that input exits 0, and 195,839 executions of a 30-second run
  complete with no crash.

primary_contact is left as-is; it is the original author's address and is
almost certainly stale. A committer on the fork should probably take it over.

+++ projects/libtsm/libtsm_fuzzer.c
@@ -18,7 +18,7 @@ static void terminal_write_fn(struct tsm_vte *vte,
     out[len % sizeof(out)] = u8[len];
 }
 
-static int term_draw_cell(struct tsm_screen *screen, uint32_t id,
+static int term_draw_cell(struct tsm_screen *screen, uint64_t id,
                           const uint32_t *ch, size_t len,
                           unsigned int cwidth, unsigned int posx,
                           unsigned int posy,
